Kolia ne spi

+++ ClientSocket/Helpers/ArrayHelper.cs
@@ -4,7 +4,7 @@ namespace ClientSocket.Helpers
 {
     static public class ArrayHelper
     {
-        public static T[] SubArray<T>(this T[] data, int index, int length)
+        public static T[] SubArray<T>(this T[] data, int index, long length)
         {
             T[] result = new T[length];
             Array.Copy(data, index, result, 0, length);
